VMR_MSL39_v1: add isolate_refseq, and QC vs taxonomy_node
QC columns compare "genome_composition" and "genome_molecule" with the values in taxonomy node. Also, changes to species in last MSL are flagged, along with proposal names.

In this release, SQL to update genbank and refseq accession numbers is generated by some python and shell scripts in https://github.com/ICTV-Virus-Knowledgebase/VMR_to_BlastDB, then run and saved here, instead of updating the [vmr] table from a separate lookup table [gb-rs]. We may return to that design in the future, but QC is much easier in python than in SQL.